% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_large_xml_venc}
\alias{prep_large_xml_venc}
\title{Transformação de grandes arquivos XML do tema LicitacaoVencedor obtidos via função 'scraper_licitacao' em data frames.}
\usage{
prep_large_xml_venc(ano = 2013, cod_mun_ibge = "411520", path_xml, path_df)
}
\arguments{
\item{ano}{ano da licitação/contrato. Default: 2013.}

\item{path_xml}{local de armazenamento dos arquivos XML.}

\item{path_df}{local para salvar a base de dados.}

\item{cod_ibge_mun}{códigos dos municípios do PR. Default: "411520".}

\item{encoding}{codificação do arquivo. Default: "latin1".}
}
\description{
Transformação de grandes arquivos XML do tema LicitacaoVencedor obtidos via função 'scraper_licitacao' em data frames.
}
\examples{
prep_large_xml_venc(ano = 2013, cod_ibge_mun = 411520, path_xml = "C:/Licitacao",
                    path_df = "C:/Licitacao")

}
