% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_xml_licitacao}
\alias{prep_xml_licitacao}
\title{Transformação de arquivos XML obtidos via função 'scraper_licitacao' em data frames.}
\usage{
prep_xml_licitacao(ano = c(2013:2017), tema = c("Licitacao",
  "LicitacaoVencedor", "LicitacaoParticipante", "Contrato", "ContratoAditivo",
  "Relacionamentos"), cod_ibge_mun = "Todos", path_xml, size_xml = 500,
  path_df, file_name = paste(ifelse(length(ano) > 1, c(paste(ano[1],
  ano[length(ano)], sep = "_")), ano[1]), tolower(tema), Sys.Date(), sep = "_"))
}
\arguments{
\item{ano}{ano da licitação/contrato. Default: 2013 a 2017.}

\item{tema}{Licitacao, LicitacaoVencedor, LicitacaoParticipante, Contrato, ContratoAditivo ou Relacionamentos.}

\item{cod_ibge_mun}{códigos dos municípios do PR. Default: "Todos".}

\item{path_xml}{local de armazenamento dos arquivos XML.}

\item{size_xml}{tamanho máximo de arquivo XML a ser considerado na tranformação (em MB). Como default foi setado o
valor de 500MB. É possível aumentar esse valor, entretanto pode sobrecarregar o processamento e ocorrer timeout.}

\item{path_df}{local para salvar a base de dados e o arquivo de erros.}

\item{file_name}{nome da base de dados a ser salva.}
}
\description{
Transforma um conjunto de arquivos XML em data frame e salva os resultados em uma base de dados. Além disso,
é salvo um arquivo de possíveis erros na tranformação. Obs.: Arquivos que excedem 500MB são desconsiderados
do processo. Assim, é necessário utilizar o arquivo de saída relativo aos 'erros' para identificá-los e
realizar o procedimento de transformação de XML em data frame separadamente!
}
\examples{
prep_xml_licitacao(ano = c(2013, 2014), tema = "Licitacao",
                   cod_ibge_mun = 410010, path_xml = "C:/Licitacao",
                   path_df = "C:/Licitacao")

}
