% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scraper_licitacao}
\alias{scraper_licitacao}
\title{Scraper para extrair dados do Portal de Informações para Todos (PIT) - Dados Abertos - do Tribunal
de Contas do Estado do Paraná (TCE/PR).}
\usage{
scraper_licitacao(ano = c(2013:2017), tema = c("Licitacao", "Contrato",
  "Relacionamentos"), cod_ibge_mun = "Todos", path)
}
\arguments{
\item{ano}{ano da licitação/contrato. Default: 2013 a 2017.}

\item{tema}{Licitacao, Contrato ou Relacionamentos.}

\item{cod_ibge_mun}{códigos dos municípios do PR. Default: "Todos".}

\item{path}{caminho para salvar os arquivos.}
}
\description{
Recebe um ano, um tema relacionado a licitações e contratos, os códigos IBGE dos municípios do PR desejados
e um caminho para salvar os arquivos. Retorna arquivos XML individuais por ano e município.
}
\examples{
scraper_licitacao(ano = c(2013, 2014), tema = "Licitacao",
                  cod_ibge_mun = 410010, path = "C:/")

}
